refactoring util
