+++ projects/01_fyyur/starter_code/util.py
@@ -1,8 +1,10 @@
-from models import db
+from datetime import datetime
+import sqlalchemy.sql.sqltypes
 import inspect
+import dateutil.parser
+import babel
 
-# e.g. groupClassBy(Venue, 'venues', 'city', 'state')
-def groupClassBy(dbClass, dbClassDictKeyName, *argv):
+def groupClassBy(db, dbClass, dbClassDictKeyName, *argv):
   listOfAttributes = []
   for arg in argv:
     listOfAttributes.append(getattr(dbClass, arg))
@@ -25,24 +27,12 @@ def groupClassBy(dbClass, dbClassDictKeyName, *argv):
 
   return groupedClassDict_withClass
 
-def populateClassObject(className, request):
-  obj = className()
-  attributes = inspect.getmembers(className, lambda a:not(inspect.isroutine(a)))
-  attributes = [a[0] for a in attributes if not(a[0].startswith('_'))]
-  for a in attributes:
-    if a in request.form:
-      if isinstance(getattr(className, a).type, db.ARRAY):
-        setattr(obj, a, request.form.getlist(a))
-      else:
-        setattr(obj, a, request.form[a])
-  return obj
-
 def populateObjectFromRequest(obj, request):
   attributes = inspect.getmembers(obj, lambda a:not(inspect.isroutine(a)))
   attributes = [a[0] for a in attributes if not(a[0].startswith('_'))]
   for a in attributes:
     if a in request.form:
-      if isinstance(getattr(type(obj), a).type, db.ARRAY):
+      if isinstance(getattr(type(obj), a).type, sqlalchemy.sql.sqltypes.ARRAY):
         setattr(obj, a, request.form.getlist(a))
       else:
         setattr(obj, a, request.form[a])
@@ -55,3 +45,18 @@ def setFormDefaultValues(form, obj):
     if hasattr(obj, a):
       setattr(getattr(form, a), "default", getattr(obj, a))
   form.process()
+
+def format_datetime(value, format='medium'):  
+  date = None
+  if type(value) is str:
+    date = dateutil.parser.parse(value)
+  elif type(value) is datetime:
+    date = value
+  else:
+    return "Error: invalid input to format_datetime"
+
+  if format == 'full':
+      format="EEEE MMMM, d, y 'at' h:mma"
+  elif format == 'medium':
+      format="EE MM, dd, y h:mma"
+  return babel.dates.format_datetime(date, format)
